+++ src/featureProvider/featureProviderSymbols.ts
@@ -1,5 +1,10 @@
 /**
  * Added to ol.Feature, if they are not part of a layer, but provided by an {@link AbstractFeatureProvider}.
  */
-// eslint-disable-next-line import/prefer-default-export
 export const isProvidedFeature: unique symbol = Symbol('isProvidedFeature');
+
+/**
+ * Added to ol.Feature, if a {@link AbstractFeatureProvider} provides more than one feature for one location.
+ * The provided feature is a cluster feature. The single features can be accessed by `feature.get('features')`.
+ */
+export const isProvidedClusterFeature = Symbol('isProvidedClusterFeature');